+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+- Changed labels for automatic ticket handling.
+- Cleaned up context menu code.
+
 ## 1.1.4 [2017-06-20]
 ### Added
 - Button to popup to quickly bring up the Service Catalog.