+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 1.1.5 [2017-07-25]
 ### Added
 - Ability to export build list to CSV file has been added.
 - Option to persist notes even after submitting (disabled by default).